README.md
The project is an application that aims to categorize individuals based on their living conditions, level of expenses, and income. It performs a multicriteria analysis considering these variables in four subgroups, based on the distribution of their income within these groups. The goal is for the individual to visualize their income distribution.